@@ -12,7 +12,7 @@ const features = [
     imageUrl: 'https://kubernetes.io/images/nav_logo.svg',
     description: (
       <>
-        Kilo can be installed on any Kubernetes cluster, allowing nodes located in different clouds or in different coutries to form a single cluster.
+        Kilo can be installed on any Kubernetes cluster, allowing nodes located in different clouds or in different countries to form a single cluster.
       </>
     ),
     clip: true,
